+++ src/components/NotFound.tsx
@@ -1,0 +1,5 @@
+import { Box } from "@mui/system";
+
+export const NotFound = () => {
+  return <Box>Not Found</Box>;
+};
