+++ src/hooks/useForm.ts
@@ -4,8 +4,8 @@ import { ChangeEvent, useState } from "react";
 export const useForm = () => {
   const [form, setForm] = useState<{ [key: string]: any }>({});
 
-  const initializeForm  = <T extends Record<string, any>> (initialValues: T) => {
-    setForm(prevValues => ({ ...prevValues, ...initialValues }));
+  const initializeForm = <T extends Record<string, any>>(initialValues: T) => {
+    setForm((prevValues) => ({ ...prevValues, ...initialValues }));
   };
 
   const handleInputChange = ({ target }: ChangeEvent<HTMLInputElement>) => {
@@ -25,16 +25,11 @@ export const useForm = () => {
     }));
   };
 
-  const resetForm =  <T extends Record<string, any>> (init: T) => {
-    setForm(init);
-  };
-
   return {
     ...form,
     form,
     handleInputChange,
     handleSelectChange,
-    resetForm,
-    initializeForm
+    initializeForm,
   };
 };
